requirements.txt
# ============================================
# Test Case Generation Agent - Dependencies
# ============================================
# Install:       pip install -r requirements.txt
# With dev/test: pip install -r requirements-dev.txt
# ============================================

# --- Core Framework ---
streamlit>=1.30.0

# --- LLM Integration ---
requests>=2.31.0
huggingface-hub>=0.20.0

# --- Document Processing ---
pypdf>=3.17.0
python-docx>=1.1.0

# --- Export ---
openpyxl>=3.1.2
pandas>=2.0.0

# --- Utilities ---
python-dotenv>=1.0.0
pydantic>=2.0.0

# --- Optional: Online LLM Providers (uncomment as needed) ---
# openai>=1.0.0
# anthropic>=0.18.0
# groq>=0.4.0

# --- Optional: Local HuggingFace Models (uncomment if running models locally) ---
# transformers>=4.36.0
# torch>=2.0.0
# accelerate>=0.25.0

# --- Optional: vLLM for high-performance local inference (uncomment for faster local models) ---
# vllm>=0.6.0
# Note: vLLM requires CUDA and works best on Linux with NVIDIA GPUs
# For installation: pip install vllm
# For CPU or non-CUDA systems, use Ollama or HuggingFace API instead
